+++ oshi-core-java25/src/main/java/oshi/driver/windows/perfmon/ProcessInformationFFM.java
@@ -5,13 +5,18 @@
 package oshi.driver.windows.perfmon;
 
 import static oshi.driver.common.windows.perfmon.PerfmonConstants.PROCESS;
+import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERFPROC_PROCESS_WHERE_IDPROCESS_0;
 import static oshi.driver.common.windows.perfmon.PerfmonConstants.WIN32_PERFPROC_PROCESS_WHERE_NAME_TOTAL;
 
+import java.util.List;
 import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.common.windows.perfmon.ProcessInformation.HandleCountProperty;
+import oshi.driver.common.windows.perfmon.ProcessInformation.IdleProcessorTimeProperty;
 import oshi.util.platform.windows.PerfCounterQueryFFM;
+import oshi.util.platform.windows.PerfCounterWildcardQueryFFM;
+import oshi.util.tuples.Pair;
 
 /**
  * Utility to query Process Information performance counter using FFM.
@@ -30,5 +35,15 @@ public final class ProcessInformationFFM {
     public static Map<HandleCountProperty, Long> queryHandles() {
         return PerfCounterQueryFFM.queryValues(HandleCountProperty.class, PROCESS,
                 WIN32_PERFPROC_PROCESS_WHERE_NAME_TOTAL);
+    }
+
+    /**
+     * Returns raw idle process performance counters.
+     *
+     * @return Raw performance counters for idle process.
+     */
+    public static Pair<List<String>, Map<IdleProcessorTimeProperty, List<Long>>> queryIdleProcessCounters() {
+        return PerfCounterWildcardQueryFFM.queryInstancesAndValues(IdleProcessorTimeProperty.class, PROCESS,
+                WIN32_PERFPROC_PROCESS_WHERE_IDPROCESS_0);
     }
 }